requirements.txt
pandas==2.2.2
numpy==1.26.4
lightgbm==4.3.0
scikit-learn==1.5.1
joblib==1.4.2
duckdb==1.1.1
mlxtend==0.23.1
streamlit==1.38.0
packaging<25
pillow<11
fastapi==0.112.0
uvicorn[standard]==0.30.6
matplotlib==3.9.2
seaborn>=0.12.0
psutil>=5.8.0
requests==2.32.3
tqdm==4.66.5
python-dotenv==1.0.1
PyYAML==6.0.2
catboost==1.2.5
xgboost==2.1.1
optuna==4.5.0
shap==0.45.1
prometheus-client==0.20.0
skl2onnx==1.19.1
